+++ src/i18next.defaults.js
@@ -49,5 +49,7 @@ var o = {
     cookieDomain: undefined,
 
     postProcess: undefined,
-    parseMissingKey: undefined
+    parseMissingKey: undefined,
+
+    shortcutFunction: 'sprintf' // or: defaultValue
 };