+++ src/i18next.js
@@ -8,7 +8,8 @@
       , resStore = {}
       , currentLng
       , replacementCounter = 0
-      , languages = [];
+      , languages = []
+      , initialized = false;
 
 
     //= i18next.exports.js